+++ packages/contracts/deploy/new/40_finalize-managing-dao/30_install-multisig-on-managing-dao.ts
@@ -9,7 +9,7 @@ import {hashHelpers} from '../../../utils/psp';
 import {MultisigSetup__factory, Multisig__factory} from '../../../typechain';
 import {HardhatRuntimeEnvironment} from 'hardhat/types';
 import {InstallationPreparedEvent} from '../../../typechain/PluginSetupProcessor';
-import {getNamedTypesFromABI} from '../../../test/plugins/utils/build-metadata';
+import {getNamedTypesFromMetadata} from '../../../utils/metadata';
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const {ethers, network} = hre;
@@ -69,7 +69,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
 
   // Prepare multisig plugin for managingDAO
   const data = ethers.utils.defaultAbiCoder.encode(
-    getNamedTypesFromABI(
+    getNamedTypesFromMetadata(
       buildMetadataJson.pluginSetup.prepareInstallation.inputs
     ),
     [approvers, [listedOnly, minApprovals]]
