+++ packages/contracts/test/plugins/governance/admin/admin-setup.ts
@@ -7,7 +7,7 @@ import {deployNewDAO} from '../../../test-utils/dao';
 import {getInterfaceID} from '../../../test-utils/interfaces';
 import metadata from '../../../../src/plugins/governance/admin/build-metadata.json';
 import {adminInterface} from './admin';
-import {getNamedTypesFromABI} from '../../utils/build-metadata';
+import {getNamedTypesFromMetadata} from '../../../../utils/metadata';
 
 const abiCoder = ethers.utils.defaultAbiCoder;
 const AddressZero = ethers.constants.AddressZero;
@@ -33,7 +33,9 @@ describe('AdminSetup', function () {
     targetDao = await deployNewDAO(ownerAddress);
 
     minimum_data = abiCoder.encode(
-      getNamedTypesFromABI(metadata.pluginSetup.prepareInstallation.inputs),
+      getNamedTypesFromMetadata(
+        metadata.pluginSetup.prepareInstallation.inputs
+      ),
       [ownerAddress]
     );
 
@@ -78,7 +80,9 @@ describe('AdminSetup', function () {
 
     it('reverts if encoded address in `_data` is zero', async () => {
       const dataWithAddressZero = abiCoder.encode(
-        getNamedTypesFromABI(metadata.pluginSetup.prepareInstallation.inputs),
+        getNamedTypesFromMetadata(
+          metadata.pluginSetup.prepareInstallation.inputs
+        ),
         [AddressZero]
       );
