+++ packages/contracts/test/plugins/governance/multisig/multisig-setup.ts
@@ -28,7 +28,7 @@ import {
   UpdatePreparedEvent,
 } from '../../../../typechain/PluginSetupProcessor';
 import {hashHelpers} from '../../../../utils/psp';
-import {getNamedTypesFromABI} from '../../utils/build-metadata';
+import {getNamedTypesFromMetadata} from '../../../../utils/metadata';
 
 const abiCoder = ethers.utils.defaultAbiCoder;
 const AddressZero = ethers.constants.AddressZero;
@@ -63,7 +63,9 @@ describe('MultisigSetup', function () {
     };
 
     minimum_data = abiCoder.encode(
-      getNamedTypesFromABI(metadata.pluginSetup.prepareInstallation.inputs),
+      getNamedTypesFromMetadata(
+        metadata.pluginSetup.prepareInstallation.inputs
+      ),
       [[signers[0].address], Object.values(defaultMultisigSettings)]
     );
 
@@ -112,7 +114,9 @@ describe('MultisigSetup', function () {
       const noMembers: string[] = [];
 
       const wrongPrepareInstallationData = abiCoder.encode(
-        getNamedTypesFromABI(metadata.pluginSetup.prepareInstallation.inputs),
+        getNamedTypesFromMetadata(
+          metadata.pluginSetup.prepareInstallation.inputs
+        ),
         [noMembers, defaultMultisigSettings]
       );
 
@@ -144,7 +148,9 @@ describe('MultisigSetup', function () {
       const members = [signers[0].address];
 
       const wrongPrepareInstallationData = abiCoder.encode(
-        getNamedTypesFromABI(metadata.pluginSetup.prepareInstallation.inputs),
+        getNamedTypesFromMetadata(
+          metadata.pluginSetup.prepareInstallation.inputs
+        ),
         [members, multisigSettings]
       );
 
@@ -175,7 +181,9 @@ describe('MultisigSetup', function () {
       const members = [signers[0].address];
 
       const wrongPrepareInstallationData = abiCoder.encode(
-        getNamedTypesFromABI(metadata.pluginSetup.prepareInstallation.inputs),
+        getNamedTypesFromMetadata(
+          metadata.pluginSetup.prepareInstallation.inputs
+        ),
         [members, multisigSettings]
       );
