+++ packages/contracts/tsconfig.json
@@ -15,6 +15,7 @@
     "./test",
     "./typechain",
     "./deploy",
-    "types/hardhat.d.ts"
+    "types/hardhat.d.ts",
+    "utils/metadata.ts"
   ]
 }
